Support bumping nodejs and gradle projects
For nodejs, the bump is performed using "npm version" command.

For gradle, the bump is performed in an oppinionated manner, expecting a propKey "version" in a gradle.properties file

For custom types,  the bump is performed reading a VERSION.txt file, which is possible to override with an environment variable. This file contains the Semver version value.